server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
langchain==0.3.0
langchain-openai==0.2.0
langchain-community==0.3.0
openai==1.50.0
pydantic==2.9.0
faiss-cpu==1.13.2
python-dotenv==1.0.1
aiosqlite==0.20.0
httpx==0.27.0
